+++ .eslintrc.js
@@ -23,6 +23,7 @@ module.exports = {
 
     'no-console': ['warn', { allow: ['warn', 'error'] }],
 
+    'import/prefer-default-export': ['off'],
     'import/no-dynamic-require': ['off'], // Dynamic requires are used to reference images.
     'global-require': ['off'], // Local requires are used to reference images.
     'react/jsx-props-no-spreading': ['off'], // I don't mind passing properties.