+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import MainView from './components/views/MainView';
 
-import { useStateStored } from './components/Preferences';
+import { useStateStored } from './components/preferences/Preferences';
 
 import './App.css';
 import './NerdFonts.css';