+++ src/components/controls/MapControlOptions.js
@@ -4,12 +4,17 @@ import clsx from 'clsx';
 import ReactSwitch from 'react-switch';
 import ReactSlider from 'rc-slider';
 import 'rc-slider/assets/index.css';
-import { resetLocalStorage, exportDataJSON, exportLegacyDataJSON } from '../Preferences';
+import { resetLocalStorage } from '../preferences/Preferences';
+import { importDataFromString } from '../preferences/DataImport';
+import { importLegacyDataFromString } from '../preferences/LegacyImport';
 import ClearMapDataPopup from '../popups/ClearMapDataPopup';
 import ExportDataPopup from '../popups/ExportDataPopup';
+import ImportDataPopup from '../popups/ImportDataPopup';
 
 import './MapControlOptions.css';
 import { t } from '../Localization';
+import { exportDataJSON } from '../preferences/DataExport';
+import { exportLegacyDataJSON } from '../preferences/LegacyExport';
 
 const MapControlOptions = ({ mapPreferences, setMapPreferences }) => {
   return (
@@ -61,6 +66,15 @@ const MapControlOptions = ({ mapPreferences, setMapPreferences }) => {
       </div>
       <div className={clsx('map-controls-options-container')}>
         <div className={clsx('map-controls-option')}>
+          <span className={clsx('map-controls-option-label')}>{t('option-import-new')}</span>
+          <ImportDataPopup
+            title={t('option-import-new')}
+            content={t('popup-import-new-content')}
+            trigger={<button type="button">{t('import')}</button>}
+            onConfirm={(data) => importDataFromString(data, setMapPreferences)}
+          />
+        </div>
+        <div className={clsx('map-controls-option')}>
           <span className={clsx('map-controls-option-label')}>{t('option-export-new')}</span>
           <ExportDataPopup
             title={t('option-export-new')}
@@ -75,6 +89,15 @@ const MapControlOptions = ({ mapPreferences, setMapPreferences }) => {
           <ClearMapDataPopup
             trigger={<button type="button">{t('clear')}</button>}
             onConfirm={resetLocalStorage}
+          />
+        </div>
+        <div className={clsx('map-controls-option')}>
+          <span className={clsx('map-controls-option-label')}>{t('option-import-old')}</span>
+          <ImportDataPopup
+            title={t('option-import-old')}
+            content={t('popup-import-old-content')}
+            trigger={<button type="button">{t('import')}</button>}
+            onConfirm={(data) => importLegacyDataFromString(data, setMapPreferences)}
           />
         </div>
         <div className={clsx('map-controls-option')}>