+++ src/components/popups/ExportDataPopup.js
@@ -5,7 +5,7 @@ import Popup from 'reactjs-popup';
 import './ExportDataPopup.css';
 import CopyTextArea from './CopyTextArea';
 
-const ExportDataPopup = ({ header, message, fetchData, trigger }) => {
+const ExportDataPopup = ({ title, message, fetchData, trigger }) => {
   const data = fetchData();
 
   // NOTE: className on popup gets overridden with <class>-content, -overlay, and -arrow.
@@ -19,7 +19,7 @@ const ExportDataPopup = ({ header, message, fetchData, trigger }) => {
       closeOnEscape
     >
       <div className={clsx('popup-export-data-container')}>
-        <span className={clsx('popup-export-data-header')}>{header}</span>
+        <span className={clsx('popup-export-data-header')}>{title}</span>
         <span className={clsx('popup-export-data-content')}>{message}</span>
         <CopyTextArea className={clsx('popup-export-data-textarea')} text={data} />
       </div>