+++ CHANGELOG.md
@@ -36,3 +36,22 @@ First major release of the rework, featuring a new interface developed in React.
   - This tab displays a progress bar for all features which are currently displayed and have at least one marker complete.
   - It includes an options menu, allowing users to pan to a random uncompleted marker, clear all markers of the feature, or clear the markers which are expired (if the respawn time is configured for that feature).
 - Added missing localization for popups.
+
+# 0.4.5
+
+- Reimplemented popups on routes.
+- Reworked the Preferences handling.
+  - Before this point I've been kinda dumb and making breaking changes to the data format without updating the vesrion, mostly assuming nobody uses the app. This updates locks in GenshinMap Data Version 1. Any future changes to the structure of the DEFAULT_MAP_PREFERENCES must now require updating DataImport.
+- Added new data importer. You can now transfer data by exporting a string and importing it.
+- Added legacy data importer. You can now transfer data from Yuanshen.site, including chest data!
+  - If you were using GenshinMap before the rework, press the Export Legacy Data button, copy that, then paste into the Import Legacy Data dialog.
+- Added data migration. If the data format used by the app has changed, it will be imported.
+  - If there is an error during this process, a value will be stored in local storage. Recovery functions can be added later as needed.
+- New features (currently empty):
+  - mondstadtLuminescentSpine
+  - liyueLuminescentSpine
+  - mondstadtCrystalCore
+  - liyueCrystalCore
+  - liyueBambooShoot
+  - liyueGoldenLoach
+- Renamed Wei to [Unusual Hillichurl](https://genshin.mihoyo.com/en/news/detail/6526).